manager/requirements.txt
fastapi
uvicorn
requests
crewai
python-dotenv